feat(ui): replace util service cconverter methods with pipes
