+++ ui/src/app/core/helpers/util.service.ts
@@ -217,24 +217,6 @@ export class UtilService
     return passedTime > milisecondsToPass ? true : false;
   }
 
-  customFormatDate(date?: string)
-  {
-    //Format from backend YYYY-MM-DD
-    if(date)
-    {
-      const day = date.split('-')[2];
-      const month = date.split('-')[1];
-      const year = date.split('-')[0];
-      return [day, month, year].join('/');
-    }
-    return date;
-  }
-
-  getTypeNameImagePath(typeIconPath: string)
-  {
-    return typeIconPath ? typeIconPath.replace("ix", "ix_names") : typeIconPath;
-  }
-
   isNaN(value)
   {
     return typeof +value !== "number" || isNaN(+value)
@@ -253,22 +235,6 @@ export class UtilService
     return undefined
   }
 
-  formatCount(count: number)
-  {
-    let formated: string = count.toString();
-    let rounded: number = count;
-    let letter: string = "";
-    const decimals: number = 2;
-    if(count >= 1e3) 
-    {
-      letter = 'K'
-      rounded = count / 1e3
-      formated = rounded.toFixed(decimals)
-    }
-    formated = formated + letter;
-    return formated;
-  }
-
   tagToChip(tag: Tag): QueryItem
   {
     return {
@@ -281,10 +247,5 @@ export class UtilService
   openInNewTab(url) 
   {
     window.open(url, '_blank');
-  }
-
-  getFlagIconUrl(countryCode?: string)
-  {
-    return countryCode ? `${this.url}images/flags/${countryCode}.svg` : countryCode;
   }
 }
